Sun, Nov 20, 2022, 5:30 PM +00:00

+++ filter-less/helpers.c
@@ -3,9 +3,9 @@
 // Convert image to grayscale
 void grayscale(int height, int width, RGBTRIPLE image[height][width])
 {
-    for (int i = 0; i < height; i++)
+    for (float i = 0; i < height; i++)
     {
-        for (int j = 0; j < width; j++)
+        for (float j = 0; j < width; j++)
         {
             float red = image[i][j].rgbtRed;
             float blue = image[i][j].rgbtBlue;
